requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
black==22.12.0
branca==0.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.88.0
flake8==6.0.0
folium==0.13.0
frozenlist==1.3.3
GeoAlchemy2==0.12.5
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
isort==5.10.1
Jinja2==3.1.2
load-dotenv==0.1.0
Mako==1.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.3
mypy-extensions==0.4.3
numpy==1.23.5
packaging==22.0
pandas==1.5.2
pathspec==0.10.3
platformdirs==2.6.0
pycodestyle==2.10.0
pydantic==1.10.2
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
requests==2.28.1
rfc3986==1.5.0
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
yarl==1.8.2
